+++ test-web/src/components/DataListView/constants/headCells.js
@@ -1,0 +1,59 @@
+// 테이블 헤더 CELL
+const headCells = [
+    {
+      id: 'trackingNo',
+      align: 'center',
+      disablePadding: false,
+      label: 'No.',
+    },
+    {
+      id: 'meatId',
+      align: 'center',
+      disablePadding: true,
+      label: '관리번호',
+    },
+    {
+      id: 'farmAddr',
+      align: 'center',
+      disablePadding: true,
+      label: '농장주소',
+    },
+    {
+      id: 'userName',
+      align: 'center',
+      disablePadding: true,
+      label: '등록인',
+    },
+    {
+      id: 'userType',
+      align: 'center',
+      disablePadding: true,
+      label: '등록인 타입',
+    },
+    {
+      id: 'company',
+      align: 'center',
+      disablePadding: true,
+      label: '소속',
+    },
+    {
+      id: 'createdAt',
+      align: 'center',
+      disablePadding: true,
+      label: '생성 날짜',
+    },
+    {
+      id: 'accept',
+      align: 'center',
+      disablePadding: false,
+      label: '승인 여부',
+    },
+    {
+      id: 'delete',
+      align: 'center',
+      disablePadding: false,
+      label: '삭제',
+    },
+  ];
+
+  export default headCells
