+++ test-web/src/components/DataListView/style/dataliststyle.js
@@ -1,0 +1,27 @@
+// DataListStyles.js
+
+const style = {
+    tableDelHeader: {
+      display: 'flex',
+      alignItems: 'center',
+      padding: '6px',
+      fontWeight: '600',
+      color: 'black',
+    },
+    tableHeader: {
+      display: 'flex',
+      alignItems: 'center',
+      padding: '6px',
+      fontWeight: '600',
+      color: '#90a4ae',
+    },
+    tableCell: {
+      align: 'left',
+      fontSize: '17px',
+      fontWeight: '600',
+      padding: '5px',
+    },
+  };
+  
+  export default style;
+
